Usability of aspects ratio in database and image structs
We have a number of functions setting the image's aspect ratio, it's content is mainly
used for collections looking for it, another code place is culling where we use the aspect
for determining how images are placed.

When calculating the aspect ratio there were rounding errors, imprecise calculations
based on preview pipe dimension or sensors requiring some cropping in rawprepare, this
leads to undesired values for aspect_ratio.
Another problem has been reported in #19252 where clearly different aspects could lead to
the same value in the database (0.7 both for 3:2 or 4:3)

This commit implements `float dt_usable_aspect(const float img_aspect)`

It returns a more stable aspect by checking for an aspect being in a +-0.5% range of
predefined aspects either provided by common sensor dimensions or the crop module.
If the tested aspects doesn't fit into one of these it returns values in 0.05 steps for portraits
and 0.1 for landscapes to avoid a possibly "exploding" list of values in collections filter.
All ratio-setting functions use above.
Also we test with with a resolution of 0.01 in all filters for matches.

Also: some code refactoring, code style modifications and some floats instead of doubles as
not required.

+++ src/views/darkroom.c
@@ -996,8 +996,8 @@ static void _dev_change_image(dt_develop_t *dev,
   if(dev->preview_pipe->backbuf
      && dev->preview_pipe->status == DT_DEV_PIXELPIPE_VALID)
   {
-    const double aspect_ratio =
-      (double)dev->preview_pipe->backbuf_width / (double)dev->preview_pipe->backbuf_height;
+    const float aspect_ratio =
+      (float)dev->preview_pipe->backbuf_width / (float)dev->preview_pipe->backbuf_height;
     dt_image_set_aspect_ratio_to(dev->preview_pipe->image.id, aspect_ratio, TRUE);
   }
   else
@@ -3168,8 +3168,8 @@ void leave(dt_view_t *self)
   // update aspect ratio
   if(dev->preview_pipe->backbuf && dev->preview_pipe->status == DT_DEV_PIXELPIPE_VALID)
   {
-    const double aspect_ratio =
-      (double)dev->preview_pipe->backbuf_width / (double)dev->preview_pipe->backbuf_height;
+    const float aspect_ratio =
+      (float)dev->preview_pipe->backbuf_width / (float)dev->preview_pipe->backbuf_height;
     dt_image_set_aspect_ratio_to(dev->preview_pipe->image.id, aspect_ratio, FALSE);
   }
   else
